requirements.txt
pocketflow>=0.0.2
pyyaml>=6.0
streamlit==1.44.1
google-cloud-aiplatform>=1.25.0
google-genai>=1.9.0